setup actualizado
- se agrego conficionales para verificaion si algun componente
 esta instado y no descargar nuevamente

+++ Make/setup.py
@@ -4,6 +4,21 @@ import urllib.request
 import zipfile
 import winreg
 
+def check_firefox_installed():
+    try:
+        # Abrir la clave del registro de Firefox
+        key = winreg.OpenKey(winreg.HKEY_LOCAL_MACHINE, r"Software\Mozilla\Mozilla Firefox")
+        
+        # Obtener el valor de la versión de Firefox
+        value, _ = winreg.QueryValueEx(key, "CurrentVersion")
+        
+        # Cerrar la clave del registro
+        winreg.CloseKey(key)
+        
+        return True
+    except FileNotFoundError:
+        return False
+
 def verify(ruta,geckodriver_url,geckodriver_zip):
     os.chdir(ruta)
     urllib.request.urlretrieve(geckodriver_url, geckodriver_zip)
@@ -12,46 +27,66 @@ def verify(ruta,geckodriver_url,geckodriver_zip):
     os.remove(geckodriver_zip)
 
 def install_firefox():
-    print("[-] Instalando Firefox...")
-    firefox_url = "https://ftp.mozilla.org/pub/firefox/releases/124.0.2/win64/es-ES/Firefox%20Setup%20124.0.2.exe"
-    firefox_installer = "Firefox Setup 124.0.2.exe"
-    urllib.request.urlretrieve(firefox_url, firefox_installer)
-    subprocess.run([firefox_installer, "-ms"])
-    os.remove(firefox_installer)
-    print("[-] Firefox instalado correctamente.")
+    if check_firefox_installed():
+        print("[-] Firefox ya instalado")
+    else:
+        print("[-]Firefox no está instalado en tu sistema.")
+        print("[-] Instalando Firefox...")
+        firefox_url = "https://ftp.mozilla.org/pub/firefox/releases/124.0.2/win64/es-ES/Firefox%20Setup%20124.0.2.exe"
+        firefox_installer = "Firefox Setup 124.0.2.exe"
+        urllib.request.urlretrieve(firefox_url, firefox_installer)
+        subprocess.run([firefox_installer, "-ms"])
+        os.remove(firefox_installer)
+        print("[-] Firefox instalado correctamente.")
 
 def install_geckodriver():
-    print("[-] Descargando geckodriver...")
-    geckodriver_url = "https://github.com/mozilla/geckodriver/releases/download/v0.33.0/geckodriver-v0.33.0-win64.zip"
-    geckodriver_zip = "geckodriver.zip"
 
-    ruta = os.path.join(os.getcwd(),'geckodriver')    
+    cmd = 'geckodriver --version'
+    msg = subprocess.run(cmd,stdout= subprocess.PIPE,stderr= subprocess.PIPE)
+    reg = msg.stdout.decode()
+    if reg.startswith('geckodriver'):
+        print('[-] Geckodriver ya Instalado')
 
-    if os.path.exists(ruta):
-        verify(ruta,geckodriver_url,geckodriver_zip)
     else:
-        os.mkdir(ruta)
-        verify(ruta,geckodriver_url,geckodriver_zip)
+        print("[-] Descargando geckodriver...")
+        geckodriver_url = "https://github.com/mozilla/geckodriver/releases/download/v0.33.0/geckodriver-v0.33.0-win64.zip"
+        geckodriver_zip = "geckodriver.zip"
 
-    print("[-] Agregando geckodriver a las variables de entorno")
+        ruta = os.path.join(os.getcwd(),'geckodriver')    
 
-    geckodriver_path = os.path.abspath("geckodriver.exe")
+        if os.path.exists(ruta):
+            verify(ruta,geckodriver_url,geckodriver_zip)
+        else:
+            os.mkdir(ruta)
+            verify(ruta,geckodriver_url,geckodriver_zip)
 
-    key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, "Environment", 0, winreg.KEY_ALL_ACCESS)
-    path_value = winreg.QueryValueEx(key, "Path")[0]
-    new_path_value = path_value + ";" + os.path.dirname(geckodriver_path)
-    winreg.SetValueEx(key, "Path", 0, winreg.REG_EXPAND_SZ, new_path_value)
-    winreg.CloseKey(key)
-    print("[-] geckodriver agregado correctamente a las variables de entorno.")
+        print("[-] Agregando geckodriver a las variables de entorno")
+
+        geckodriver_path = os.path.abspath("geckodriver.exe")
+
+        key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, "Environment", 0, winreg.KEY_ALL_ACCESS)
+        path_value = winreg.QueryValueEx(key, "Path")[0]
+        new_path_value = path_value + ";" + os.path.dirname(geckodriver_path)
+        winreg.SetValueEx(key, "Path", 0, winreg.REG_EXPAND_SZ, new_path_value)
+        winreg.CloseKey(key)
+        print("[-] geckodriver agregado correctamente a las variables de entorno.")
 
 def install_python():
-    print("[-] Instalando Python...")
-    python_url = "https://www.python.org/ftp/python/3.12.1/python-3.12.1-amd64.exe"
-    python_installer = "python-3.12.1-amd64.exe"
-    urllib.request.urlretrieve(python_url, python_installer)
-    subprocess.run([python_installer, "/quiet", "InstallAllUsers=1", "PrependPath=1"])
-    os.remove(python_installer)
-    print("[-] Python instalado correctamente.")
+    cmd = 'python --version'
+    msg = subprocess.run(cmd,stdout= subprocess.PIPE,stderr= subprocess.PIPE)
+    reg = msg.stdout.decode()
+
+    if reg.startswith('Python 3.12.1'):
+        print('[-] python ya instado')
+
+    else:
+        print("[-] Instalando Python...")
+        python_url = "https://www.python.org/ftp/python/3.12.1/python-3.12.1-amd64.exe"
+        python_installer = "python-3.12.1-amd64.exe"
+        urllib.request.urlretrieve(python_url, python_installer)
+        subprocess.run([python_installer, "/quiet", "InstallAllUsers=1", "PrependPath=1"])
+        os.remove(python_installer)
+        print("[-] Python instalado correctamente.")
 
 def create_virtual_environment():
     print("[-] Creando entorno virtual...")
@@ -67,7 +102,8 @@ def activate_virtual_environment():
     subprocess.run([screenshots_path])
     print("[-] Entorno virtual activado.")
 
-    subprocess.run(["pip", "install", "selenium", "pillow", "playwright"])
+    subprocess.run(["pip", "install", "selenium", "pillow", "playwright"],stdout=subprocess.PIPE,stderr=subprocess.PIPE,stdin=subprocess.PIPE)
+    subprocess.run(["playwright","install"])
     print("[-] Librerías instaladas correctamente.")
 
     print("[-] Descargando Scripts...")
@@ -79,7 +115,7 @@ def activate_virtual_environment():
     urllib.request.urlretrieve(GuiRemoto, GuiRemoto[-12:])
 
     print("[-] Scripts Descargados ...")
-    exec('python GuiRemoto.py')
+    os.system('python GuiRemoto.py')
 
 def setup():
     install_python()
